+++ wechatkit/exceptions.py
@@ -1,18 +1,18 @@
 """Wechatkit exception module."""
 
 
-class WechatkitBaseException(Exception):
-    """Wechatkit Exception."""
+class WechatKitBaseException(Exception):
+    """Wechatkit base Exception."""
 
     def __init__(self, error_info):
         """Init."""
-        super(WechatkitBaseException, self).__init__(error_info)
+        super(WechatKitBaseException, self).__init__(error_info)
         self.error_info = error_info
 
 
-class WechatException(WechatkitBaseException):
+class WechatKitException(WechatKitBaseException):
     """Wechatkit Exception."""
 
 
-class WechatSignException(WechatException):
+class WechatSignException(WechatKitException):
     """Wechat Sign Exception."""